HapiSchemaMigrationTest intermittent failure fix
