+++ hapi-fhir-jpaserver-test-utilities/src/main/java/ca/uhn/fhir/jpa/embedded/OracleCondition.java
@@ -19,8 +19,7 @@
  */
 package ca.uhn.fhir.jpa.embedded;
 
-import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.SystemUtils;
+import ca.uhn.fhir.jpa.util.DatabaseSupportUtil;
 import org.junit.jupiter.api.extension.ConditionEvaluationResult;
 import org.junit.jupiter.api.extension.ExecutionCondition;
 import org.junit.jupiter.api.extension.ExtensionContext;
@@ -33,25 +32,8 @@ public class OracleCondition implements ExecutionCondition {
 
 	@Override
 	public ConditionEvaluationResult evaluateExecutionCondition(ExtensionContext theExtensionContext) {
-		return canUseOracle()
+		return DatabaseSupportUtil.canUseOracle()
 				? ConditionEvaluationResult.enabled(ENABLED_MSG)
 				: ConditionEvaluationResult.disabled(DISABLED_MSG);
-	}
-
-	public static boolean canUseOracle() {
-		if (!isMac()) {
-			return true;
-		}
-		return isColimaConfigured();
-	}
-
-	private static boolean isMac() {
-		return SystemUtils.IS_OS_MAC || SystemUtils.IS_OS_MAC_OSX;
-	}
-
-	private static boolean isColimaConfigured() {
-		return StringUtils.isNotBlank(System.getenv("TESTCONTAINERS_DOCKER_SOCKET_OVERRIDE"))
-				&& StringUtils.isNotBlank(System.getenv("DOCKER_HOST"))
-				&& System.getenv("DOCKER_HOST").contains("colima");
 	}
 }
